+++ resources/views/app/components/iconLabel.blade.php
@@ -11,7 +11,7 @@
             </span>
         @else
             <span class="w-5 flex justify-center">
-                <i class="{{ $icon ?? 'far fa-arrow-right' }} {{ $caution ?? null ? 'text-red-500' : '' }}"></i>
+                <i class="{{ $icon ?? 'fas fa-arrow-right' }} {{ $caution ?? null ? 'text-red-500' : '' }}"></i>
             </span>
         @endisset
     @endif
@@ -27,7 +27,7 @@
                 <span class="counter ml-2">{{ $count }} </span>
             </span>
         @else
-            <i class="{{ $icon ?? 'far fa-arrow-right' }} {{ $caution ?? null ? 'text-red-500' : '' }}"></i>
+            <i class="{{ $icon ?? 'fas fa-arrow-right' }} {{ $caution ?? null ? 'text-red-500' : '' }}"></i>
         @endisset
     @endif
 </span>