+++ src/Domain/Audience/Actions/Subscribers/ConfirmSubscriberAction.php
@@ -4,6 +4,7 @@ namespace Spatie\Mailcoach\Domain\Audience\Actions\Subscribers;
 
 use Spatie\Mailcoach\Domain\Audience\Actions\Subscribers\Concerns\SendsWelcomeMail;
 use Spatie\Mailcoach\Domain\Audience\Events\SubscribedEvent;
+use Spatie\Mailcoach\Domain\Audience\Events\TagAddedEvent;
 use Spatie\Mailcoach\Domain\Audience\Models\Subscriber;
 
 class ConfirmSubscriberAction
@@ -27,6 +28,10 @@ class ConfirmSubscriberAction
 
         if ($this->sendWelcomeMail) {
             $this->sendWelcomeMail($subscriber);
+        }
+
+        foreach ($subscriber->tags as $tag) {
+            event(new TagAddedEvent($subscriber, $tag));
         }
 
         event(new SubscribedEvent($subscriber));